refactor(cli): decompose cli.js dispatch into scripts/cli/ modules (CORE-1)
Extract help text and the eleven DAG/coordinator case blocks from
scripts/cli.js into scripts/cli/ modules, following the existing
ratify-command.js + shared.js pattern:

- scripts/cli/help.js — printHelp() usage text (verbatim move)
- scripts/cli/spec-resolution.js — resolveSpecId, isAmbiguousSpec,
  requireSpecId, resolveMergeOrCleanupSpec (+ private epic→spec index)
- scripts/cli/dag-commands.js — status/next/complete/block/parallel/
  expand/merge/cleanup/sync/reboot/loop handlers behind runDagCommand()

cli.js keeps argument parsing and a thin dispatch switch only.

Acceptance evidence (plan docs/plans/2026-06-11-capability-seam-fix-
implementation-plan.md #### CORE-1):
- wc -l scripts/cli.js = 198 (<= 600; headroom 502 vs 700 hard limit)
- npm test: all 5 runners green (exit 0); npm run lint: green
- Byte-identical CLI behavior: 50-invocation before/after battery
  (help, schema, all coordinator commands incl. multi-spec aggregate/
  ambiguity/error paths, expand/cleanup with real worktrees) — empty diff
- tests/node/test-cli.js passes with zero assertion changes
  (git diff --name-only -- tests/ is empty)
